Move some shared code to utils
